+++ polaris-react/src/components/ActionList/ActionList.tsx
@@ -47,6 +47,10 @@ export function ActionList({
     finalSections = sections;
   }
 
+  // const isFilterable = finalSections?.some((section) =>
+  //   section.items.some((item) => typeof item.content === 'string'),
+  // );
+
   const hasMultipleSections = finalSections.length > 1;
   const elementRole =
     hasMultipleSections && actionRole === 'menuitem' ? 'menu' : undefined;