+++ polaris-tokens/scripts/toStyleSheet.ts
@@ -1,91 +1,78 @@
 import fs from 'fs';
 import path from 'path';
 
-import type {Metadata, MetadataGroup} from '../src';
-import type {MetadataBase} from '../src/types';
+import type {MetaThemeShape, MetaTokenGroupShape} from '../src/themes/types';
+import {metaThemeVariantPartials, metaThemeDefault} from '../src/themes';
+import {themeNameDefault} from '../src/themes/constants';
+import {createThemeSelector} from '../src/themes/utils';
+import {createVar} from '../src/utilities';
+import type {Entries} from '../src/types';
 
 const cssOutputDir = path.join(__dirname, '../dist/css');
-const sassOutputDir = path.join(__dirname, '../dist/scss');
+const scssOutputDir = path.join(__dirname, '../dist/scss');
 const cssOutputPath = path.join(cssOutputDir, 'styles.css');
-const sassOutputPath = path.join(sassOutputDir, 'styles.scss');
+const scssOutputPath = path.join(scssOutputDir, 'styles.scss');
 
-/**
- * Creates static CSS custom properties.
- * Note: These values don't vary by color-scheme.
- */
-export function getStaticCustomProperties(metadata: Metadata) {
-  return Object.entries(metadata)
-    .map(([_, tokenGroup]) => getCustomProperties(tokenGroup))
+/** Creates CSS declarations from a base or variant partial theme. */
+export function getMetaThemeDecls(metaTheme: MetaThemeShape) {
+  return Object.values(metaTheme)
+    .map((metaTokenGroup) => getMetaTokenGroupDecls(metaTokenGroup))
     .join('');
 }
 
-/**
- * Creates static CSS custom properties overrides.
- * Note: These values don't vary by color-scheme.
- */
-export function getStaticCustomPropertiesExperimental(metadata: MetadataBase) {
-  return Object.entries(metadata)
-    .map(([_, tokenGroup]) =>
-      getCustomProperties(
-        Object.fromEntries(
-          Object.entries(tokenGroup)
-            // Only include tokens with `valueExperimental` prop
-            .filter(([_, metadataProperties]) =>
-              Boolean(metadataProperties.valueExperimental),
-            )
-            // Move `valueExperimental` to `value` position
-            .map(([tokenName, metadataProperties]) => [
-              tokenName,
-              {value: metadataProperties.valueExperimental!},
-            ]),
-        ),
-      ),
+/** Creates CSS declarations from a token group. */
+export function getMetaTokenGroupDecls(metaTokenGroup: MetaTokenGroupShape) {
+  return Object.entries(metaTokenGroup)
+    .map(([tokenName, {value}]) =>
+      tokenName.startsWith('motion-keyframes')
+        ? `${createVar(tokenName)}:p-${tokenName};`
+        : `${createVar(tokenName)}:${value};`,
     )
     .join('');
 }
 
-/**
- * Creates CSS custom properties for a given metadata object.
- */
-export function getCustomProperties(tokenGroup: MetadataGroup) {
-  return Object.entries(tokenGroup)
-    .map(([token, {value}]) =>
-      token.startsWith('motion-keyframes') || token.startsWith('keyframes')
-        ? `--p-${token}:p-${token};`
-        : `--p-${token}:${value};`,
-    )
-    .join('');
-}
-
-/**
- * Concatenates the `keyframes` token-group into a single string.
- */
-export function getKeyframes(motion: MetadataGroup) {
+/** Creates `@keyframes` rules for `motion-keyframes-*` tokens. */
+export function getKeyframes(motion: MetaTokenGroupShape) {
   return Object.entries(motion)
-    .filter(
-      ([token]) =>
-        token.startsWith('motion-keyframes') || token.startsWith('keyframes'),
-    )
-    .map(([token, {value}]) => `@keyframes p-${token}${value}`)
+    .filter(([tokenName]) => tokenName.startsWith('motion-keyframes'))
+    .map(([tokenName, {value}]) => `@keyframes p-${tokenName}${value}`)
     .join('');
 }
 
-export async function toStyleSheet(metadata: Metadata) {
-  if (!fs.existsSync(cssOutputDir)) {
-    await fs.promises.mkdir(cssOutputDir, {recursive: true});
-  }
-  if (!fs.existsSync(sassOutputDir)) {
-    await fs.promises.mkdir(sassOutputDir, {recursive: true});
-  }
+export async function toStyleSheet() {
+  await fs.promises.mkdir(cssOutputDir, {recursive: true}).catch((error) => {
+    if (error.code !== 'EEXIST') {
+      throw error;
+    }
+  });
 
-  const styles = `
-  :root{color-scheme:light;${getStaticCustomProperties(metadata)}}
-  html.Polaris-Summer-Editions-2023{${getStaticCustomPropertiesExperimental(
-    metadata,
-  )}}
-  ${getKeyframes(metadata.motion)}
-`;
+  await fs.promises.mkdir(scssOutputDir, {recursive: true}).catch((error) => {
+    if (error.code !== 'EEXIST') {
+      throw error;
+    }
+  });
+
+  const metaThemeVariantPartialsEntries = Object.entries(
+    metaThemeVariantPartials,
+  ).filter(([themeName]) => themeName !== themeNameDefault) as Entries<
+    Omit<typeof metaThemeVariantPartials, typeof themeNameDefault>
+  >;
+
+  const styles = [
+    `:root{color-scheme:light;${getMetaThemeDecls(metaThemeDefault)}}`,
+    metaThemeVariantPartialsEntries.map(
+      ([themeName, metaThemeVariantPartial]) =>
+        `${createThemeSelector(themeName)}{${getMetaThemeDecls(
+          metaThemeVariantPartial,
+        )}}`,
+    ),
+    getKeyframes(metaThemeDefault.motion),
+    // Newline terminator
+    '',
+  ]
+    .flat()
+    .join('\n');
 
   await fs.promises.writeFile(cssOutputPath, styles);
-  await fs.promises.writeFile(sassOutputPath, styles);
+  await fs.promises.writeFile(scssOutputPath, styles);
 }