+++ polaris-tokens/src/themes/utils.ts
@@ -1,0 +1,57 @@
+import deepmerge from 'deepmerge';
+
+import type {Entry, Exact} from '../types';
+import {createExact} from '../utilities';
+
+import type {
+  ExtractMetaThemeValues,
+  ExtractMetaTokenGroupValues,
+  MetaThemeShape,
+  MetaThemeVariant,
+  MetaThemeVariantPartialShape,
+  MetaTokenGroupShape,
+  ThemeName,
+} from './types';
+import {themeNameLightUplift} from './constants';
+import {metaThemeBase} from './base';
+
+export const createMetaThemeVariantPartial =
+  createExact<MetaThemeVariantPartialShape>();
+
+export function createMetaThemeVariant<
+  T extends Exact<MetaThemeVariantPartialShape, T>,
+>(metaThemeVariantPartial: T): MetaThemeVariant {
+  return deepmerge(metaThemeBase, metaThemeVariantPartial);
+}
+
+export function createThemeClassName(themeName: ThemeName) {
+  return themeName === themeNameLightUplift
+    ? themeName
+    : `p-theme-${themeName}`;
+}
+
+export function createThemeSelector(themeName: ThemeName) {
+  return `html.${createThemeClassName(themeName)}`;
+}
+
+export function extractMetaTokenGroupValues<T extends MetaTokenGroupShape>(
+  metaTokenGroup: T,
+) {
+  return Object.fromEntries(
+    Object.entries(metaTokenGroup).map(
+      ([tokenName, {value}]): Entry<
+        ExtractMetaTokenGroupValues<MetaTokenGroupShape>
+      > => [tokenName, value],
+    ),
+  ) as ExtractMetaTokenGroupValues<T>;
+}
+
+export function extractMetaThemeValues<T extends MetaThemeShape>(metaTheme: T) {
+  return Object.fromEntries(
+    Object.entries(metaTheme).map(
+      ([tokenGroupName, metaTokenGroup]): Entry<
+        ExtractMetaThemeValues<MetaThemeShape>
+      > => [tokenGroupName, extractMetaTokenGroupValues(metaTokenGroup)],
+    ),
+  ) as ExtractMetaThemeValues<T>;
+}